Move to miner directory
